REST API Authorization with JSON Web Tokens
This introduces another auth mechanism for authentication to to REST
endpoints: using self-contained JSON Web Tokens (RFC 7519) containing
pre-agreed claims related to the identity and the access granted by an
external trusted identity/authorization provider.

With this approach, user and role management as well as the authentication
mechanism is handled by the external IdP and the information in the token
is simply used as-is because of the trust relationship between ESH and the
IdP service. Of course, this currently breaks existing UIs and other clients
of the API because they don't know how to request a token from the IdP.
Ultimately, an OpenID Connect or OAuth2 server delivering this type of
tokens should be embedded within ESH itself.

JWTs may also contain scopes limiting the access of the client application
(relying party) to certain operations - for instance, some apps should
not be allowed access to endpoints managing things or bindings, even if the
user (subject) himself has the admin role, because they have no business
doing so. The RequiredScopesDynamicFeature ensures the scopes required by
the API operation (for now, described in the Swagger specification) are
found in the provided token. In practice, the IdP service will require the
consent of the user the first time with a dedicated screen ("This
application will be able to: manage your things, update your items, etc."),
but this could be bypassed for "trusted" clients like the standard UIs.

The tokens are supposed to be signed using JWS HS256 (the only one
supported for now) and their signature will be verified against a
Base64-encoded pre-shared secret defined in the service configuration - the
issuer claim ('iss') is also expected to match the value specified in the
configuration.

Signed-off-by: Yannick Schaus <eclipse@schaus.net>